refactor : 강의 후기(Review)를 모두 DB에서 관리 및 조회, 그에 따른 Lectures 반환 로직 수정 #187

+++ src/main/java/inu/codin/codin/domain/lecture/entity/LectureEntity.java
@@ -23,10 +23,10 @@ public class LectureEntity {
     private List<String> semester;
 
     private double starRating;
-    private long participants;
+    private int participants;
     private Emotion emotion;
 
-    public void updateReviewRating(double starRating, long participants, Emotion emotion){
+    public void updateReviewRating(double starRating, int participants, Emotion emotion){
         this.starRating = starRating;
         this.participants = participants;
         this.emotion = emotion;
